requirements.txt
# Dependências principais
boto3==1.34.34
pandas==2.2.0
pyarrow==15.0.0
awswrangler==3.6.0

# Validação de dados
great-expectations==0.18.8
pydantic==2.6.0

# Utilitários
python-dotenv==1.0.1
pyyaml==6.0.1

# AWS
aws-lambda-powertools==2.31.0

# Testes
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
moto==5.0.0

# Qualidade de código
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Registro de logs
structlog==24.1.0
